test: introduce unit and integration testing for Node strategy
